creacion de notificaciones automaticas

+++ db/migrate/20160912000734_create_notifications.rb
@@ -2,7 +2,7 @@ class CreateNotifications < ActiveRecord::Migration
 	def change
 		create_table :notifications do |t|
 			t.integer :from_type
-			t.integer :from_id
+			t.bigint :from_id
 			t.text :message
 			t.boolean :readed
